fix #772, optimize drawing of events in MonthView
Informal proof that it's optimal: A greedy algorithm is used.  Assume
the first N events have been drawn in an optimal way.  The next event
draw is the one which has start-date as close as possible to the
previeous' end-date, without overlapping.  (When a line is filled up,
the next is begun.)  If there are any "unused" days between these two
events, they could not have been used by any other remaining event
anyway.  The length of the event doesen't matter (so we can always
choose the longest one, of the ones with the same start-date) because
every event has to be draw, so the remaining events can always be
arranged in an optimal way.  This means that the new event was placed in
an optimal way.