Update strings.xml
I translated the following lines to German:
13
14
88
95
105
106
110
185
186
187
188
189
190
191
205
241
243
245
247
249
255-272
274
276

The only German terms missing here, are in lines 107-109, because it is unclear in which context the terms "Going", "Not going" and "Maybe going" are being used. Since that part is about attendees, the term "going" could mean something like "possible" or "leaving". Translating the wrong meaning could end up sounding rather ridiculous.

Anka