Always shift end date by one day for all-day events
Shift the end date by one day when the start and end date are equal.
Previously this was only done when the end date is already ahead of the
start date.  An event with equal start and end date is created when the
start and end times are equal before switching to all-day.