Resolves (or at least improves): - #1538 - https://github.com/bitfireAT/davx5-ose/discussions/66 - https://github.com/Kozea/Radicale/issues/1264 - May also resolve some of the issues described in #1983.
The "EXDATE" property is only meant to be set on events with the "RRULE" property. See discussion in #1538, especially the response from DAVx5 support listed in the first comment.

I tried also setting STATUS_CANCELED on these recurrence exception events, but DAVx5 would sometimes no longer recognize them as belonging to the parent event and would give the child events different UIDs, which is undesirable. It seems like it works great to just leave out the EXDATE.

I have not tested this patch with other CalDAV setups, but it solves the problem for me on DAVx5+Radicale.