Properly handle `Update all occurrences` option
Closes https://github.com/SimpleMobileTools/Simple-Calendar/issues/981

The events were moved forward because the start and end timestamps from the repeating occurrence were saved instead of the original start and end times.

The call to `eventsHelper.addEventRepeatLimit()` was removed because I believe it was added there by mistake and the changes were overwritten anyways due to the next call to `eventsHelper.updateEvent()`. Another reason to remove it is to avoid triggering a CalDAV update which may interfere with the event update.