Use local timezone to configure weekly views
This addresses issues like https://github.com/SimpleMobileTools/Simple-Calendar/issues/2143

The weekly view originally used local timezone to configure everything but was modified to use UTC to fix a `Start week with Sunday` issue https://github.com/SimpleMobileTools/Simple-Calendar/issues/1213 where events were created on the wrong day. Since `getFirstDayOfWeek()` was rewritten to handle week preferences properly, there is no need to complicate things using UTC anymore (otherwise we'll have to handle UTC-to-local-timezone conversion in many places)

This also fixes an issue with the drag-and-drop feature where the events were moved by ±1 day because the difference in timezone was not taken care of.